+++ CHANGES.rst
@@ -25,6 +25,7 @@ New Features
   - Automatically use ``guess=False`` when reading if the file ``format`` is
     provided and the format parameters are uniquely specified.  This update
     also removes duplicate format guesses to improve performance. [#3418]
+- ``astropy.table``
 
   - Calls to ascii.read() for fixed-width tables may now omit one of the keyword
     arguments ``col_starts`` or ``col_ends``. Columns will be assumed to begin and
@@ -189,6 +190,10 @@ Other Changes and Additions
 
 - Nothing yet.
 
+1.0.1 (unreleased)
+------------------
+  - Added ``Table.to_pandas`` and ``Table.from_pandas`` for converting to/from
+    pandas dataframes.
 
 1.0.3 (unreleased)
 ------------------